Renderer rework
- new AbstractRenderer to shield to colorer from the concrete renderers
- new AbstractGraphics2DRenderer that draws to any given Graphics2D
- EngineWord no longer public
- ProcessingSketchRenderer is now much shorter, just the setting of the
graphics, the dimensions, and the ignoring of the close
- SVGDocumentRenderer is now much shorter, just the setting of the
graphics, bogous default dimension and the writing of the stream after
the drawing is done
