+++ src/wordcram/AbstractRenderer.java
@@ -1,0 +1,12 @@
+package wordcram;
+
+public abstract class AbstractRenderer implements Renderer {
+	
+	WordColorer colorer;
+
+	public void setColorer(WordColorer colorer) {
+		this.colorer = colorer;
+	}
+
+	
+}
